feat: implement TypeScript support and refactor MIDI parsing logic
