+++ index.html
@@ -49,6 +49,6 @@
     <p style="color: #888; font-size: 0.9em; padding-left: 10px;">Coming Soon...</p>
   </div>
 
-<script type="module" src="js/app.js"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
